Closes #60. Split classes into files and put images into one separate folder.
