+++ tests/images/smooth0.png
