+++ tests/images/smooth100.png
