+++ tests/images/smooth50.png
