+++ tests/images/test_brightness_darken.png
