+++ tests/images/test_brightness_lighten.png
