+++ tests/images/test_high_contrast.png
