+++ tests/images/test_hist_eq0.png
