+++ tests/images/test_hist_eq10.png
