+++ tests/images/test_hist_eq5.png
