+++ tests/images/test_low_contrast.png
